Refactor: move Category.java to data.models
